backend/app/utils/db.py
import psycopg2
from psycopg2.extras import RealDictCursor
from flask import current_app

def get_db_connection():
    return psycopg2.connect(
        host=current_app.config["DB_HOST"],
        port=current_app.config["DB_PORT"],
        dbname=current_app.config["DB_NAME"],
        user=current_app.config["DB_USER"],
        password=current_app.config["DB_PASSWORD"],
        cursor_factory=RealDictCursor,
        connect_timeout=5
    )
