backend/requirements.txt
flask
flask-jwt-extended
psycopg2-binary
python-dotenv
gunicorn
